+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Remove `update-cli-pins` job from `release.yaml`. Renovate already handles updating `gha-utils` version.
+
 ## [5.7.2 (2026-02-11)](https://github.com/kdeldycke/workflows/compare/v5.7.1...v5.7.2)
 
 - Replace `taiki-e/install-action` with direct `curl` download for `labelmaker` binary.